Move code related to fetching data from database to api folder
