+++ src/modules/services/hooks/useCompanyServices.ts
@@ -1,23 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
-import { collection, query, getDocs, where } from "firebase/firestore";
-import { db } from "firebase-config";
-import { parseGetDocs } from "common/utils/firebaseHelpers";
-import { Service } from "modules/services/types";
+import { getCompanyServices } from "../api";
 
 const useCompanyServices = (companyId: string, initFetch = true) => {
-  const servicesCollectionRef = collection(db, "services");
-
   const { data, ...queryResult } = useQuery(
     ["services", companyId],
-    async () => {
-      const q = query(
-        servicesCollectionRef,
-        where("companyId", "==", companyId)
-      );
-      const data = await getDocs(q);
-
-      return parseGetDocs<Service[]>(data);
-    },
+    () => getCompanyServices(companyId),
     { enabled: initFetch }
   );
   return { ...queryResult, services: data };
